fix(coinjoin): offer the shielded destination from every move-funds surface
The destination-choice sheet (Dash Wallet vs Shielded) was reachable only from
the post-sync popup, and any dismissal of that popup — the close button or a
swipe — is persisted as "Later". Re-arming it needs the CoinJoin balance to rise
above the dismissed amount, and with mixing removed from the app that balance can
only fall, so one swipe left the direct CoinJoin -> Shielded route unreachable for
good. The durable Settings and Tools rows called sweepCoinJoin() directly, which
always moves the coins into the transparent balance.

The Settings and Tools rows now branch exactly as the popup does, and the
availability test they share moved out of HomeViewModel into
CoinJoinMoveDestinationPolicy so the surfaces cannot drift apart on what the user
is offered. Neither menu surface persists a "Later" on dismissal: they are
user-initiated, so closing the sheet means "not now", not "never ask again".

Also splits the 460-line CoinJoinMoveFundsSheet.swift into one type per file
under CoinJoinMoveFunds/ and adds previews for all seven stages, which the flow
badly needed — reproducing it live requires a wallet with a CoinJoin balance.
