+++ DashWallet/Sources/UI/Menu/Tools/ToolsMenuViewModel.swift
@@ -36,6 +36,11 @@ class ToolsMenuViewModel: ObservableObject {
     @Published var csvExportData: (fileName: String, file: URL)?
     @Published var safariLink: String?
     @Published var showCoinJoinSweepConfirmation = false
+    /// Destination-choice sheet (Dash Wallet vs Shielded). Presented
+    /// instead of the plain confirmation when the shielded route is
+    /// available, so this row offers the same choice as the post-sync
+    /// popup rather than silently sweeping to the transparent balance.
+    @Published var showCoinJoinMoveFundsSheet = false
     @Published var coinJoinSweepErrorMessage: String?
     /// True while the diagnostic-log archive is being staged + zipped
     /// (file I/O off the main actor). Guards re-entry from row taps.
@@ -52,7 +57,7 @@ class ToolsMenuViewModel: ObservableObject {
     private static let minCoinJoinSweepDuffs: UInt64 = 1000
 
     /// Live CoinJoin-account spendable balance (duffs) — the SDK source of truth.
-    private var coinJoinLeftoverDuffs: UInt64 {
+    var coinJoinLeftoverDuffs: UInt64 {
         SwiftDashSDKWalletState.shared.coinJoinBalanceDuffs
     }
 
@@ -152,7 +157,18 @@ class ToolsMenuViewModel: ObservableObject {
                     subtitle: NSLocalizedString("CoinJoin is no longer supported", comment: "CoinJoin"),
                     icon: .custom("image.coinjoin.menu", maxHeight: 22),
                     action: { [weak self] in
-                        self?.showCoinJoinSweepConfirmation = true
+                        guard let self else { return }
+                        // Same branch as the post-sync popup
+                        // (`HomeViewModel.maybeShowCoinJoinSweepDialog`): offer the
+                        // destination choice whenever the shielded route is viable,
+                        // and fall back to the transparent-only confirmation when
+                        // it is not.
+                        if CoinJoinMoveDestinationPolicy.shieldedDestinationAvailable(
+                            forBalanceDuffs: self.coinJoinLeftoverDuffs) {
+                            self.showCoinJoinMoveFundsSheet = true
+                        } else {
+                            self.showCoinJoinSweepConfirmation = true
+                        }
                     }
                 ),
                 at: storageIdx + 1
